Moves definition of few structs into a new file
